Last warning fix

+++ src/expression_evaluator.cpp
@@ -344,10 +344,8 @@ InternalValue CallExpression::CallGlobalRange(RenderContext& values)
     auto items_count = distance / step;
     items_count = items_count < 0 ? 0 : static_cast<size_t>(items_count);
 
-    return ListAdapter::CreateAdapter(items_count, [start, step](size_t idx) {
-        return InternalValue(static_cast<int64_t>(start + step * idx));
-    });
-
+    return ListAdapter::CreateAdapter(static_cast<size_t>(items_count),
+                                      [start, step](size_t idx) { return InternalValue(static_cast<int64_t>(start + step * idx)); });
 }
 
 InternalValue CallExpression::CallLoopCycle(RenderContext& values)
